requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.7
python-telegram-bot==20.5
XlsxWriter==3.2.0
